add torrent via cloud
