+++ node/lib/newUpload.js
@@ -6,7 +6,7 @@ import Debug from 'debug'
 import sanitize from 'sanitize-filename'
 import { dialog, ipcMain } from 'electron'
 import { getMainWindow } from './window'
-import { serverGetAsync, isCloud, uploadTorrent } from './server'
+import { serverGetAsync, isCloud } from './server'
 import { createTask } from './uploadTransform'
 
 Promise.promisifyAll(fs) // babel would transform Promise to bluebird
@@ -165,27 +165,6 @@ const uploadHandle = (event, args) => {
   })
 }
 
-const addTorrentHandle = (event, args) => {
-  const { dirUUID } = args
-  const filters = [
-    { name: 'Torrent', extensions: ['torrent'] },
-    { name: 'All Files', extensions: ['*'] }
-  ]
-  dialog.showOpenDialog(getMainWindow(), { properties: ['openFile'], filters }, (entries) => {
-    if (!entries || !entries.length) return
-    uploadTorrent(dirUUID, entries[0], (err, res) => {
-      if (err) {
-        console.log('addTorrent error', err.response && err.response.body)
-        let text = i18n.__('Add Torrent Failed')
-        if (err.response && err.response.body && err.response.body.message === 'torrent exist') text = i18n.__('Task Exist')
-        getMainWindow().webContents.send('snackbarMessage', { message: text })
-      } else {
-        getMainWindow().webContents.send('snackbarMessage', { message: i18n.__('Add Torrent Success') })
-      }
-    })
-  })
-}
-
 const dragFileHandle = (event, args) => {
   let entries = args.files
   if (!entries || !entries.length) return
@@ -221,4 +200,3 @@ ipcMain.on('UPLOADMEDIA', uploadMediaHandle)
 ipcMain.on('DRAG_FILE', dragFileHandle)
 ipcMain.on('resolveConflicts', resolveHandle)
 ipcMain.on('START_TRANSMISSION', startTransmissionHandle)
-ipcMain.on('ADD_TORRENT', addTorrentHandle)
